refactor: flatten is-blank module
