+++ src/commands/cowsay/index.ts
@@ -1,7 +1,7 @@
 import { say } from 'cowsay';
 import { type ChatInputCommandInteraction, SlashCommandBuilder, type TextChannel } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { isBlank } from '../../utils/isBlank';
+import { isBlank } from '../../utils/is-blank';
 import { logger } from '../../utils/logger';
 import { fetchLastMessageBeforeId } from '../../utils/messageFetcher';
 import type { Command } from '../builder';
