+++ src/commands/insult/index.ts
@@ -1,5 +1,5 @@
 import { type ChatInputCommandInteraction, SlashCommandBuilder } from 'discord.js';
-import { isBlank } from '../../utils/isBlank';
+import { isBlank } from '../../utils/is-blank';
 import { logger } from '../../utils/logger';
 import type { Command } from '../builder';
 import { randomInsultGenerator } from './insultGenerator';
