+++ src/commands/mockSomeone/index.ts
@@ -1,6 +1,6 @@
 import { type ChatInputCommandInteraction, SlashCommandBuilder, type TextChannel } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { isBlank } from '../../utils/isBlank';
+import { isBlank } from '../../utils/is-blank';
 import { logger } from '../../utils/logger';
 import { fetchLastMessageBeforeId } from '../../utils/messageFetcher';
 import { getRandomBoolean } from '../../utils/random';
