+++ src/commands/weather/index.ts
@@ -1,6 +1,6 @@
 import { type ChatInputCommandInteraction, SlashCommandBuilder } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { isBlank } from '../../utils/isBlank';
+import { isBlank } from '../../utils/is-blank';
 import { logger } from '../../utils/logger';
 import type { Command } from '../builder';
 import { fetchWeather } from './fetchWeather';
